refactor: move public announcement endpoints to public-announcements.ts
